trading_framework/requirements.txt
python-binance>=1.0.32
pandas>=1.3.0
numpy>=1.21.0
pyyaml>=5.4.0
